--- a/modules/apps/asset/asset-tags-service/src/main/java/com/liferay/asset/tags/internal/search/spi/model/index/contributor/AssetTagDocumentContributor.java
+++ b/modules/apps/asset/asset-tags-service/src/main/java/com/liferay/asset/tags/internal/search/spi/model/index/contributor/AssetTagDocumentContributor.java
@@ -16,6 +16,8 @@ import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentContributor;
 import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
@@ -37,6 +39,13 @@ public class AssetTagDocumentContributor
 
 	@Override
 	public void contribute(Document document, BaseModel<AssetTag> baseModel) {
+		ServiceContext serviceContext =
+			ServiceContextThreadLocal.getServiceContext();
+
+		if ((serviceContext != null) && serviceContext.isStrictAdd()) {
+			return;
+		}
+
 		String className = document.get(Field.ENTRY_CLASS_NAME);
 
 		long classNameId = portal.getClassNameId(className);