--- a/modules/apps/headless/headless-builder/headless-builder-impl/src/main/java/com/liferay/headless/builder/internal/model/listener/APIPropertyRelevantObjectEntryModelListener.java
+++ b/modules/apps/headless/headless-builder/headless-builder-impl/src/main/java/com/liferay/headless/builder/internal/model/listener/APIPropertyRelevantObjectEntryModelListener.java
@@ -22,6 +22,7 @@ import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
@@ -118,8 +119,8 @@ public class APIPropertyRelevantObjectEntryModelListener
 		try {
 			Map<String, Serializable> values = objectEntry.getValues();
 
-			long apiSchemaId = (long)values.get(
-				"r_apiSchemaToAPIProperties_c_apiSchemaId");
+			long apiSchemaId = GetterUtil.getLong(
+				values.get("r_apiSchemaToAPIProperties_c_apiSchemaId"));
 
 			if (!_validationHelper.isValidObjectEntry(
 					"L_API_SCHEMA", apiSchemaId)) {
@@ -181,8 +182,8 @@ public class APIPropertyRelevantObjectEntryModelListener
 				}
 			}
 
-			long parentAPIPropertyId = (long)values.get(
-				"r_apiPropertyToAPIProperties_c_apiPropertyId");
+			long parentAPIPropertyId = GetterUtil.getLong(
+				values.get("r_apiPropertyToAPIProperties_c_apiPropertyId"));
 
 			if (parentAPIPropertyId != 0) {
 				if (!_validationHelper.isValidObjectEntry(