--- a/modules/apps/headless/headless-builder/headless-builder-impl/src/main/java/com/liferay/headless/builder/internal/model/listener/APISchemaRelevantObjectEntryModelListener.java
+++ b/modules/apps/headless/headless-builder/headless-builder-impl/src/main/java/com/liferay/headless/builder/internal/model/listener/APISchemaRelevantObjectEntryModelListener.java
@@ -15,6 +15,7 @@ import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.io.Serializable;
@@ -58,8 +59,10 @@ public class APISchemaRelevantObjectEntryModelListener
 
 			if (!_validationHelper.isValidObjectEntry(
 					"L_API_APPLICATION",
-					(long)values.get(
-						"r_apiApplicationToAPISchemas_c_apiApplicationId"))) {
+					GetterUtil.getLong(
+						values.get(
+							"r_apiApplicationToAPISchemas_c_" +
+								"apiApplicationId")))) {
 
 				throw new ObjectEntryValuesException.InvalidObjectField(
 					null, "An API schema must be related to an API application",