--- a/modules/apps/object/object-admin-rest-api/src/main/java/com/liferay/object/admin/rest/dto/v1_0/util/ObjectFieldUtil.java
+++ b/modules/apps/object/object-admin-rest-api/src/main/java/com/liferay/object/admin/rest/dto/v1_0/util/ObjectFieldUtil.java
@@ -32,8 +32,7 @@ public class ObjectFieldUtil {
 			"defaultValue",
 			com.liferay.object.field.setting.util.ObjectFieldSettingUtil.
 				getDefaultValueAsString(
-					null, objectField.getObjectFieldId(),
-					objectFieldSettingLocalService, null)
+					null, objectField, objectFieldSettingLocalService, null)
 		).put(
 			"externalReferenceCode", objectField.getExternalReferenceCode()
 		).put(