--- a/modules/apps/object/object-api/src/main/resources/com/liferay/object/field/setting/util/packageinfo
+++ b/modules/apps/object/object-api/src/main/resources/com/liferay/object/field/setting/util/packageinfo
@@ -1,1 +1,1 @@
-version 1.3.0
+version 2.0.0