--- a/modules/apps/object/object-service/src/main/java/com/liferay/object/model/impl/ObjectEntryImpl.java
+++ b/modules/apps/object/object-service/src/main/java/com/liferay/object/model/impl/ObjectEntryImpl.java
@@ -17,7 +17,7 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.service.CompanyLocalServiceUtil;
-import com.liferay.portal.kernel.util.HashMapBuilder;
+import com.liferay.portal.kernel.util.Validator;
 
 import java.io.Serializable;
 
@@ -76,13 +76,16 @@ public class ObjectEntryImpl extends ObjectEntryBaseImpl {
 					objectDefinition.getTitleObjectFieldId());
 
 			if (objectField != null) {
+				String title = ObjectEntryValuesUtil.getValueString(
+					objectField, getValues());
+
+				if (Validator.isNotNull(title)) {
+					return title;
+				}
+
 				return ObjectEntryValuesUtil.getValueString(
 					objectField,
-					HashMapBuilder.create(
-						getValues()
-					).putAll(
-						ObjectEntryLocalServiceUtil.getSystemValues(this)
-					).build());
+					ObjectEntryLocalServiceUtil.getSystemValues(this));
 			}
 		}