--- a/modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryLocalServiceImpl.java
+++ b/modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryLocalServiceImpl.java
@@ -311,6 +311,7 @@ public class ObjectEntryLocalServiceImpl
 		ServiceContext resourcePermissionServiceContext = new ServiceContext();
 
 		resourcePermissionServiceContext.setIndexingEnabled(false);
+		resourcePermissionServiceContext.setStrictAdd(true);
 
 		_resourcePermissionLocalService.addResourcePermissions(
 			objectEntry.getCompanyId(), objectEntry.getGroupId(),