… build-test.xml
… modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectFieldSettingLocalServiceImpl.java
… modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/entries/display/context/ObjectEntryDisplayContextImpl.java
… modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/buffer/BufferedIndexerInvocationHandler.java
… modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/buffer/util/IndexerRequestBufferExecutorUtil.java
… modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/index/UpdateDocumentIndexWriterImpl.java
… modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/spi/model/index/contributor/GroupedModelDocumentContributor.java
… modules/apps/sharing/sharing-search/src/main/java/com/liferay/sharing/search/internal/permission/SharingEntrySearchPermissionDocumentContributor.java
… modules/util/portal-tools-db-upgrade-client/properties/portal-upgrade-database.properties
… portal-impl/src/com/liferay/portal/service/impl/ResourcePermissionLocalServiceImpl.java
… portal-impl/src/com/liferay/portlet/asset/service/impl/AssetEntryLocalServiceImpl.java
… portal-impl/src/com/liferay/portlet/social/service/impl/SocialActivityCounterLocalServiceImpl.java
… portal-impl/src/portal.properties
… portal-impl/test/portal-test.properties
… portal-kernel/src/com/liferay/portal/kernel/search/SearchContext.java
… portal-kernel/src/com/liferay/portal/kernel/search/StrictObjectReindexThreadLocal.java
… portal-kernel/src/com/liferay/portal/kernel/service/ServiceContext.java
… test.properties