This is a complete redo for https://liferay.atlassian.net/browse/LPD-25633

I optimized off 92.53% of the total execution time, which means it is 13.39 times faster. @victorhcunha will post detailed test results later.

This pull completely removes the linear relationship between import entry count and DB select/update query count. No matter how many entries you are importing, we only fire out a near constant number of select/update queries. Of course the insert query count still has a linear bond with entry count, after all the whole thing is about inserting the data into DB.

The tx model is untouched, we are not merging multi-tx into one.

Search index model is refactored to have batch support. Each individual index request is still fired out right after the tx commit to push the work to ES asap. Therefore we still have the same number of index requests to ES through the entire import process. But none of those requests is asking for ES index refresh. At the end of the import process, after all index requests are processed, we fire out one last refresh request to make the changes visible from search.

It is important to keep the per entry tx model, not just because of the previous deadlock issue with HSQL, but also to make sure we actually can handle large batch import. With single tx model, all changes are buffered not only at Hibernate session/JDBC driver level, but also all search index requests are buffered too. For large batch import, we will run out of heap space for that. Also due to the buffering, ES server is not getting anything to work on until the end of the tx, then all work goes to it all at once, and portal has to sync wait ES to finish in order to continue. It is very wasteful working style. Portal is the main work executor, but it heavily uses both DB and ES. To make the work load balanced across the entire system, portal should dispatch work to db and ES as evenly and smoothly as possible. In real production env, portal/db/es are going to be 3 individual hardwares, they can truly work in parallel, let's why this multi tx model can help overall throughput.


@marco-leo I waited that ObjectEntry.getValues() change from object team for a few days, nothing happened. I ended up making the change myself to optimize just enough for the test case, see https://github.com/brianchandotcom/liferay-portal/commit/b5b0a9d45ab2ab9b7ac4299da7aada23ac51fbc5 there are still 2 loose ends that can be improved, but at least they are not needed for the current test case. Object team should continue the improvement as separate tasks.
Two pending items:
1) If the ObjectEntry's ObjectFields are localized, the prepopulation of the localization values are still querying db. This can be avoid by intercepting all localization insertions for prepopulation.
2) If the ObjectEntry's ObjectFields are not statical values, means they are "Aggregation" or "Formula" typed, then the values are not statically available from inputs. Those Aggregation/Formula logic need to be applied to the input values in order to have the collect values for prepopulation.

CC @vicnate5 @wesleygong @4lejandrito @BryanEngler @gustavolimav