@brianchandotcom 2024-08-22
@shuyangzhou Great job on this PR.

Please resend on top of https://github.com/shuyangzhou/liferay-portal/pull/11605

It's too fragile right now if the good performance requires certain parameters in the JDBC URL. See my commits where I found lots of discrepancies between maria db, mysql, aurora mysql (yes, another one), and portal.properties vs. portal-test.properties vs. setup mysql.

I want to remove ALL of those parameters, and dynamically set them IF they are not in the URL.

That way, I don't have to communicate with all our implementations / GS / cloud to get the right parameters out.

Thx.