@victorhcunha 2024-08-22
Here are the detailed results:

5 rounds using one of our performance testing machines

Round | Master Baseline c62dc40 (ms) | PR (ms) | Difference (%)
-- | -- | -- | --
1 | 1699608 | 122771 | -92.77651082
2 | 1694395 | 129150 | -92.37781037
3 | 1641447 | 119132 | -92.74225729
4 | 1634282 | 122196 | -92.52295503
5 | 1610472 | 125669 | -92.19675971

The final results considering the 5 rounds:

-- | Master Baseline c62dc40  (ms) | PR  (ms) | Difference from Master (%)
-- | -- | -- | --
Average | 1656040.8 | 123783.6 | -92.52532909
Standard Deviation | 39153.81681 | 3792.630921 | -90.31350905

Based on the tests, we obtained an import time 92.53% shorter.